Updated dependency for aws-cloudformation-rpdk-java-plugin
