+++ aws-logs-loggroup/src/main/java/software/amazon/logs/loggroup/Configuration.java
@@ -1,4 +1,4 @@
-package com.aws.logs.loggroup;
+package software.amazon.logs.loggroup;
 
 import org.json.JSONObject;
 import org.json.JSONTokener;
