+++ aws-logs-loggroup/src/main/java/software/amazon/logs/loggroup/DeleteHandler.java
@@ -1,9 +1,9 @@
-package com.aws.logs.loggroup;
+package software.amazon.logs.loggroup;
 
-import com.amazonaws.cloudformation.proxy.AmazonWebServicesClientProxy;
-import com.amazonaws.cloudformation.proxy.Logger;
-import com.amazonaws.cloudformation.proxy.ProgressEvent;
-import com.amazonaws.cloudformation.proxy.ResourceHandlerRequest;
+import software.amazon.cloudformation.proxy.AmazonWebServicesClientProxy;
+import software.amazon.cloudformation.proxy.Logger;
+import software.amazon.cloudformation.proxy.ProgressEvent;
+import software.amazon.cloudformation.proxy.ResourceHandlerRequest;
 import software.amazon.awssdk.services.cloudwatchlogs.model.ResourceNotFoundException;
 
 import java.util.Objects;
@@ -22,7 +22,7 @@ public class DeleteHandler extends BaseHandler<CallbackContext> {
             proxy.injectCredentialsAndInvokeV2(Translator.translateToDeleteRequest(model),
                 ClientBuilder.getClient()::deleteLogGroup);
         } catch (final ResourceNotFoundException e) {
-            throw new com.amazonaws.cloudformation.exceptions.ResourceNotFoundException(ResourceModel.TYPE_NAME,
+            throw new software.amazon.cloudformation.exceptions.ResourceNotFoundException(ResourceModel.TYPE_NAME,
                 Objects.toString(model.getPrimaryIdentifier()));
         }
