+++ aws-logs-loggroup/src/test/java/software/amazon/logs/loggroup/DeleteHandlerTest.java
@@ -1,21 +1,22 @@
-package com.aws.logs.loggroup;
+package software.amazon.logs.loggroup;
 
-import com.amazonaws.cloudformation.exceptions.ResourceNotFoundException;
-import com.amazonaws.cloudformation.proxy.AmazonWebServicesClientProxy;
-import com.amazonaws.cloudformation.proxy.Logger;
-import com.amazonaws.cloudformation.proxy.OperationStatus;
-import com.amazonaws.cloudformation.proxy.ProgressEvent;
-import com.amazonaws.cloudformation.proxy.ResourceHandlerRequest;
+import software.amazon.cloudformation.exceptions.ResourceNotFoundException;
+import software.amazon.cloudformation.proxy.AmazonWebServicesClientProxy;
+import software.amazon.cloudformation.proxy.Logger;
+import software.amazon.cloudformation.proxy.OperationStatus;
+import software.amazon.cloudformation.proxy.ProgressEvent;
+import software.amazon.cloudformation.proxy.ResourceHandlerRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.ArgumentMatchers;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import software.amazon.awssdk.services.cloudwatchlogs.model.DeleteLogGroupResponse;
 import software.amazon.awssdk.services.cloudwatchlogs.model.DescribeLogGroupsResponse;
 import software.amazon.awssdk.services.cloudwatchlogs.model.LogGroup;
 
-import java.util.Arrays;
+import java.util.Collections;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -24,8 +25,8 @@ import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
 
 @ExtendWith(MockitoExtension.class)
-public class ReadHandlerTest {
-    ReadHandler handler;
+public class DeleteHandlerTest {
+    DeleteHandler handler;
 
     @Mock
     private AmazonWebServicesClientProxy proxy;
@@ -35,22 +36,24 @@ public class ReadHandlerTest {
 
     @BeforeEach
     public void setup() {
-        handler = new ReadHandler();
+        handler = new DeleteHandler();
         proxy = mock(AmazonWebServicesClientProxy.class);
         logger = mock(Logger.class);
     }
 
     @Test
     public void handleRequest_Success() {
+        final DeleteLogGroupResponse deleteResponse = DeleteLogGroupResponse.builder().build();
+
         final LogGroup logGroup = LogGroup.builder()
                 .logGroupName("LogGroup")
                 .retentionInDays(1)
                 .build();
         final DescribeLogGroupsResponse describeResponse = DescribeLogGroupsResponse.builder()
-                .logGroups(Arrays.asList(logGroup))
+                .logGroups(Collections.singletonList(logGroup))
                 .build();
 
-        doReturn(describeResponse)
+        doReturn(describeResponse, deleteResponse)
             .when(proxy)
             .injectCredentialsAndInvokeV2(
                 ArgumentMatchers.any(),
@@ -73,18 +76,14 @@ public class ReadHandlerTest {
         assertThat(response.getCallbackContext()).isNull();
         assertThat(response.getCallbackDelaySeconds()).isEqualTo(0);
         assertThat(response.getResourceModels()).isNull();
-        assertThat(response.getResourceModel()).isEqualToComparingFieldByField(logGroup);
+        assertThat(response.getResourceModel()).isNull();
         assertThat(response.getMessage()).isNull();
         assertThat(response.getErrorCode()).isNull();
     }
 
     @Test
-    public void handleRequest_FailureNotFound_EmptyLogGroupResponse() {
-        final DescribeLogGroupsResponse describeResponse = DescribeLogGroupsResponse.builder()
-                .logGroups(Arrays.asList())
-                .build();
-
-        doReturn(describeResponse)
+    public void handleRequest_FailureNotFound() {
+        doThrow(software.amazon.awssdk.services.cloudwatchlogs.model.ResourceNotFoundException.class)
                 .when(proxy)
                 .injectCredentialsAndInvokeV2(
                         ArgumentMatchers.any(),
@@ -99,50 +98,6 @@ public class ReadHandlerTest {
         final ResourceHandlerRequest<ResourceModel> request = ResourceHandlerRequest.<ResourceModel>builder()
                 .desiredResourceState(model)
                 .build();
-
-        assertThrows(ResourceNotFoundException.class,
-            () -> handler.handleRequest(proxy, request, null, logger));
-    }
-
-    @Test
-    public void handleRequest_FailureNotFound_WithException() {
-        doThrow(software.amazon.awssdk.services.cloudwatchlogs.model.ResourceNotFoundException.class)
-            .when(proxy)
-            .injectCredentialsAndInvokeV2(
-                ArgumentMatchers.any(),
-                ArgumentMatchers.any()
-            );
-
-        final ResourceModel model = ResourceModel.builder()
-            .logGroupName("LogGroup")
-            .retentionInDays(1)
-            .build();
-
-        final ResourceHandlerRequest<ResourceModel> request = ResourceHandlerRequest.<ResourceModel>builder()
-            .desiredResourceState(model)
-            .build();
-
-        assertThrows(ResourceNotFoundException.class,
-            () -> handler.handleRequest(proxy, request, null, logger));
-    }
-
-    @Test
-    public void handleRequest_FailureNotFound_NullLogGroupInput() {
-        final ResourceModel model = ResourceModel.builder()
-            .build();
-
-        final ResourceHandlerRequest<ResourceModel> request = ResourceHandlerRequest.<ResourceModel>builder()
-            .desiredResourceState(model)
-            .build();
-
-        assertThrows(ResourceNotFoundException.class,
-            () -> handler.handleRequest(proxy, request, null, logger));
-    }
-
-    @Test
-    public void handleRequest_FailureNotFound_NullModel() {
-        final ResourceHandlerRequest<ResourceModel> request = ResourceHandlerRequest.<ResourceModel>builder()
-            .build();
 
         assertThrows(ResourceNotFoundException.class,
             () -> handler.handleRequest(proxy, request, null, logger));
