+++ aws-logs-loggroup/src/test/java/software/amazon/logs/loggroup/UpdateHandlerTest.java
@@ -1,10 +1,10 @@
-package com.aws.logs.loggroup;
+package software.amazon.logs.loggroup;
 
-import com.amazonaws.cloudformation.proxy.AmazonWebServicesClientProxy;
-import com.amazonaws.cloudformation.proxy.Logger;
-import com.amazonaws.cloudformation.proxy.OperationStatus;
-import com.amazonaws.cloudformation.proxy.ProgressEvent;
-import com.amazonaws.cloudformation.proxy.ResourceHandlerRequest;
+import software.amazon.cloudformation.proxy.AmazonWebServicesClientProxy;
+import software.amazon.cloudformation.proxy.Logger;
+import software.amazon.cloudformation.proxy.OperationStatus;
+import software.amazon.cloudformation.proxy.ProgressEvent;
+import software.amazon.cloudformation.proxy.ResourceHandlerRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -180,7 +180,7 @@ public class UpdateHandlerTest {
             .desiredResourceState(model)
             .build();
 
-        assertThrows(com.amazonaws.cloudformation.exceptions.ResourceNotFoundException.class,
+        assertThrows(software.amazon.cloudformation.exceptions.ResourceNotFoundException.class,
             () -> handler.handleRequest(proxy, request, null, logger));
     }
 }
